requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
datasets
pypdf
python-dotenv
pandas
flask
gunicorn
prometheus_client